home/requirements.txt
coreapi==2.3.3
Django==2.1.7
django-autoslug==1.9.4
django-model-utils==3.1.2
djangorestframework==3.9.2
pytz==2018.9
PyYAML==5.1
